pyproject.toml
[tool.poetry]
name = "inbac"
version = "1.1.0"
description = "Interactive batch cropper made for quick cropping of images."
authors = ["Robert Węcławski <r.weclawski@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/weclaw1/inbac"
repository = "https://github.com/weclaw1/inbac"
keywords = ["inbac", "batch", "image", "crop"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.6",
    "License :: OSI Approved :: MIT License",
    "Topic :: Multimedia :: Graphics :: Editors",
    "Topic :: Utilities",
]
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.6"
pillow = "^7.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3"
pylint = "^2.4"
autopep8 = "^1.4"
pyinstaller = "^3.6"
macholib = "^1.14"
pywin32-ctypes = "^0.2.0"
pefile = "^2019.4.18"

[tool.poetry.scripts]
inbac = "inbac.inbac:main"
test = "tests.test_inbac:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
